updated Heartbeat and Connectivity code

+++ src/mobile/dsm/utils/Utility.java
@@ -1,6 +1,6 @@
 package mobile.dsm.utils;
 
 public class Utility {
-	public static final int HEARTBEATTIME = 5;
+	public static final long HEARTBEATTIME = 5L;
 	public static final int MB = 1024 * 1024;
 }
